duck_image/private/solver/pyproject.toml
[tool.poetry]
name = "solver"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^10.4.0"
decompyle3 = {git = "https://github.com/rocky/python-decompile3.git"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
